vunit: Install include headers into separate directory
It's better to keep the installed headers inside directory named after
the project to avoid naming conflicts.
