+++ subprojects/vinumc/tests/vec_macros.c
@@ -1,4 +1,4 @@
-#include <vunit.h>
+#include <vunit/vunit.h>
 
 #include "../vec.h"
